refactor: extract common odometry parameters to shared header
