+++ cpp/include/sycl_points/pipeline/lidar_inertial_odometry_params.hpp
@@ -2,7 +2,7 @@
 
 #include "sycl_points/algorithms/lio/lio_registration.hpp"
 #include "sycl_points/algorithms/registration/registration_params.hpp"
-#include "sycl_points/pipeline/lidar_odometry_params.hpp"
+#include "sycl_points/pipeline/odometry_common_params.hpp"
 
 namespace sycl_points {
 namespace pipeline {
@@ -10,9 +10,10 @@ namespace lidar_inertial_odometry {
 
 /// @brief Parameters for the LiDAR-Inertial Odometry pipeline.
 ///
-/// Extends lidar_odometry::Parameters with an LIO-specific optimization block.
+/// Extends the parameters shared by the odometry pipelines with an
+/// LIO-specific optimization block.
 /// IMU is always required; params_.imu.enable is forced true in the pipeline.
-struct Parameters : public lidar_odometry::Parameters {
+struct Parameters : public odometry::CommonParameters {
     struct LIO {
         /// Total number of IEKF/Gauss-Newton iterations per frame.
         /// The IEKF loop is flat: each iteration runs one linearize+solve, and the
